cleaned up a few weird lines of code, should still work the same

+++ src/drivers/motors/Motors.cpp
@@ -39,9 +39,11 @@ void Motors::Controller::spin(int power) {
 
     if (m_power > 0) {
         pwm_set_chan_level(m_in1Slice, m_in1Channel, Drivers::Motors::MAX_POWER);
-        pwm_set_chan_level(m_in2Slice, m_in2Channel, Drivers::Motors::MAX_POWER - m_power);
+        pwm_set_chan_level(m_in2Slice, m_in2Channel,
+                           Drivers::Motors::MAX_POWER - static_cast<uint16_t>(m_power));
     } else if (m_power < 0) {
-        pwm_set_chan_level(m_in1Slice, m_in1Channel, Drivers::Motors::MAX_POWER + m_power);
+        pwm_set_chan_level(m_in1Slice, m_in1Channel,
+                           Drivers::Motors::MAX_POWER + static_cast<uint16_t>(m_power));
         pwm_set_chan_level(m_in2Slice, m_in2Channel, Drivers::Motors::MAX_POWER);
     } else {
         pwm_set_chan_level(m_in1Slice, m_in1Channel, Drivers::Motors::MAX_POWER);
